backend/requirements.txt
fastapi
uvicorn
motor
pydantic-settings
pandas
dnspython
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
python-multipart
python-dotenv
pymupdf
sentence-transformers
langchain-text-splitters
joblib
numpy
scikit-learn
xgboost
pdfplumber
google-generativeai
